+++ tests/module3/monopoly_realistischTest.py
@@ -28,8 +28,8 @@ def hassimuleer_groot_aantal_potjes_Monopoly(test):
 	def testMethod():
 		correctFunction = False
 
-		if assertlib.fileContainsFunctionDefinitions(_fileName, "simuleer_groot_aantal_potjes_Monopoly"):
-			nArguments = len(lib.getFunction("simuleer_groot_aantal_potjes_Monopoly", _fileName).arguments)
+		if assertlib.fileContainsFunctionDefinitions(_fileName, "simuleer_groot_aantal_potjes_monopoly"):
+			nArguments = len(lib.getFunction("simuleer_groot_aantal_potjes_monopoly", _fileName).arguments)
 
 			if nArguments == 3:
 				correctFunction = True
@@ -40,7 +40,7 @@ def hassimuleer_groot_aantal_potjes_Monopoly(test):
 
 	test.test = testMethod
 	test.fail = lambda info : "zorg dat de functie drie argumenten heeft, het aantal potjes, startgeld voor speler 1 en startgeld voor speler 2"
-	test.description = lambda : "definieert de functie simuleer_potje en simuleer_groot_aanal_potjes_Monopoly met drie argumenten"
+	test.description = lambda : "definieert de functie simuleer_potje_monopoly en simuleer_groot_aanal_potjes_monopoly met drie argumenten"
 	test.timeout = lambda : 90
 
 
@@ -48,14 +48,14 @@ def hassimuleer_groot_aantal_potjes_Monopoly(test):
 @t.test(10)
 def correctAverageDiv(test):
 	def testMethod():
-		outcome = lib.getFunction("simuleer_groot_aantal_potjes_Monopoly", _fileName)(10000, 1500, 1500)
+		outcome = lib.getFunction("simuleer_groot_aantal_potjes_monopoly", _fileName)(10000, 1500, 1500)
 		if assertlib.sameType(outcome, None):
-			info = "Zorg er voor dat de functie simuleer_groot_aantal_potjes_Monopoly het verschil in het bezit van straten returnt en alleen deze waarde returnt"
-		elif assertlib.between(outcome, 0, 99999999):
+			info = "Zorg er voor dat de functie simuleer_groot_aantal_potjes_monopoly het verschil in het bezit van straten returnt en alleen deze waarde returnt"
+		elif assertlib.between(outcome, -99999999, 0):
 			info = "Als speler 1 meer straten heeft dan speler 2 is het verschil positief"
 		else:
 			info = "Het verschil is niet erg groot, gemiddeld zelfs minder dan 1 straat"
-		return assertlib.between(outcome, .45, .15), info
+		return assertlib.between(outcome, .15, .45), info
 
 	test.test = testMethod
 	test.description = lambda : "Monopoly met twee spelers geeft de het correcte gemiddelde verschil in gekochten straten"
